+++ languageServer/src/main/scala/core/parsers/FlatMap.scala
@@ -1,21 +1,23 @@
 package core.parsers
 
+//TODO add tests
 class FlatMap[Input <: ParseInput, +Result, +NewResult](left: Parser[Input, Result],
-                                                        f: Result => Parser[Input, NewResult])
-  extends Parser[Input, NewResult] { //TODO add tests
+                                                        getRight: Result => Parser[Input, NewResult])
+  extends Parser[Input, NewResult] {
 
   override def parseNaively(input: Input, state: ParseState): ParseResult[NewResult] = {
     val leftResult = left.parseCached(input, state)
     leftResult match {
       case leftSuccess: ParseSuccess[Result] =>
-        val right = f(leftSuccess.result)
+        val right = getRight(leftSuccess.result)
         val rightResult = right.parseCached(leftSuccess.remainder, state)
         rightResult match {
           case rightSuccess: ParseSuccess[NewResult] =>
             rightSuccess.
               addFailure(leftSuccess.biggestFailure match {
                 case NoFailure => NoFailure
-                case ParseFailure(partialResult, remainder, message) => ParseFailure(partialResult.flatMap(r => f(r).getDefault(state)), remainder, message)
+                case ParseFailure(partialResult, remainder, message) =>
+                  ParseFailure(partialResult.flatMap(leftPartial => getRight(leftPartial).getDefault(state)), remainder, message)
               })
 
           case rightFailure: ParseFailure[NewResult] =>
@@ -31,7 +33,7 @@ class FlatMap[Input <: ParseInput, +Result, +NewResult](left: Parser[Input, Resu
       case leftFailure: ParseFailure[Result] =>
         val result = for {
           leftPartial <- leftFailure.partialResult
-          rightDefault <- f(leftPartial).getDefault(state)
+          rightDefault <- getRight(leftPartial).getDefault(state)
         } yield rightDefault
         ParseFailure(result, leftFailure.remainder, leftFailure.message)
     }
@@ -39,6 +41,6 @@ class FlatMap[Input <: ParseInput, +Result, +NewResult](left: Parser[Input, Resu
 
   override def getDefault(cache: DefaultCache): Option[NewResult] = for {
     leftDefault <- cache(left)
-    rightDefault <- cache(f(leftDefault))
+    rightDefault <- cache(getRight(leftDefault))
   } yield rightDefault
 }
