+++ languageServer/src/main/scala/core/parsers/sequences/SequenceInput.scala
@@ -1,0 +1,9 @@
+package core.parsers.sequences
+
+import core.parsers.ParseInput
+
+trait SequenceInput[Input, Elem] extends ParseInput {
+  def atEnd: Boolean
+  def head: Elem
+  def tail: Input
+}
