+++ languageServer/src/main/scala/core/parsers/strings/StringReader.scala
@@ -1,6 +1,6 @@
 package core.parsers.strings
 
-import core.parsers.SequenceInput
+import core.parsers.sequences.SequenceInput
 
 import scala.util.parsing.input.OffsetPosition
